requirements.txt
anthropic>=0.89.0
pydantic>=2.0.0
python-dotenv>=1.0.0
